requirements.txt
streamlit==1.32.2

numpy==1.26.4
pillow==9.5.0
opencv-python-headless==4.8.1.78

deepface==0.0.79
mtcnn==0.1.1

tensorflow==2.13.0
keras==2.13.1

protobuf==3.20.3
h5py==3.9.0
scipy==1.10.1
scikit-learn==1.3.2
matplotlib==3.7.1
